Add support for multiple slack tokens

+++ index.js
@@ -1,7 +1,9 @@
-const itunes = require("./playback/index.js");
+// const itunes = require("./playback/index.js");
 const request = require("request");
 const encodeurl = require("encodeurl");
 const fs = require("fs");
+const Application = require("jxa").Application;
+const iTunes = Application("Music");
 
 const path = __dirname + "/shouldClear.txt";
 
@@ -9,8 +11,21 @@ require("dotenv").config({
 	path: "/Users/maxmatthews/Developer/slacktunes-node/.env"
 });
 
-itunes.currentTrack(data => {
-	if (data) {
+setInterval(() => {
+	// console.log("hit");
+	let data;
+	try {
+		const name = iTunes.currentTrack.name();
+		const artist = iTunes.currentTrack.artist();
+		const album = iTunes.currentTrack.album();
+		const state = iTunes.playerState();
+		data = { name, artist, album, state };
+	} catch (e) {
+		console.log(e);
+		data = { name: null };
+	}
+	// console.log(data.state);
+	if (data.name && data.state === "playing") {
 		if (data.artist.length > 25) {
 			data.artist = data.artist.substring(0, 20) + "...";
 		}
@@ -28,28 +43,36 @@ itunes.currentTrack(data => {
 			.replace(/\&/g, "%26")
 			.replace(/\(|\)|\?|\+|\&/g, "");
 
-		request({
-			url: `https://slack.com/api/users.profile.set?token=${process.env.SLACK_TOKEN}&profile=%7B%22status_emoji%22%3A%22:headphones:%22%2C%22status_text%22%3A%22${trackInfo}%22%7D`,
-			method: "POST"
-		}).on("response", () => {
-			if (!fs.existsSync(path)) {
-				fs.writeFileSync(path, "");
+		for (const [key, value] of Object.entries(process.env)) {
+			if (key.includes("SLACK_TOKEN")) {
+				request({
+					url: `https://slack.com/api/users.profile.set?token=${value}&profile=%7B%22status_emoji%22%3A%22:headphones:%22%2C%22status_text%22%3A%22${trackInfo}%22%7D`,
+					method: "POST"
+				}).on("response", () => {
+					if (!fs.existsSync(path)) {
+						fs.writeFileSync(path, "");
+					}
+				});
 			}
-
-			process.exit();
-		});
+		}
 	} else {
 		if (fs.existsSync(path)) {
-			request({
-				url: `https://slack.com/api/users.profile.set?token=${process.env.SLACK_TOKEN}&profile=%7B%22status_emoji%22%3A%22%22%2C%22status_text%22%3A%22%22%7D`,
-				method: "POST"
-			}).on("response", () => {
-				fs.unlinkSync(path);
+			for (const [key, value] of Object.entries(process.env)) {
+				if (key.includes("SLACK_TOKEN")) {
+					request({
+						url: `https://slack.com/api/users.profile.set?token=${
+							process.env.SLACK_TOKEN
+						}&profile=%7B%22status_emoji%22%3A%22%22%2C%22status_text%22%3A%22%22%7D`,
+						method: "POST"
+					}).on("response", () => {
+						fs.unlinkSync(path);
 
-				process.exit();
-			});
+						// process.exit();
+					});
+				}
+			}
 		} else {
-			process.exit(); //the status was already cleared, so we don't have to do it, just end the script
+			// process.exit(); //the status was already cleared, so we don't have to do it, just end the script
 		}
 	}
-});
+}, 1000 * 60);
